refactor: refactor blocs

+++ app/lib/bloc/user_account_gym_bloc/user_account_gym_bloc.dart
@@ -1,4 +1,5 @@
 import 'package:app/api/user_account_gym/user_account_gym_service.dart';
+import 'package:app/resources/app_settings.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter/material.dart';
 import 'package:hive/hive.dart';
@@ -17,44 +18,67 @@ class UserAccountGymBloc extends Bloc<UserAccountGymEvent, UserAccountGymState> 
   }
 
   void _onPostUserAccountGym(PostUserAccountGym event, Emitter<UserAccountGymState> emit) async {
+    if (!AppSettings.hasConnection) {
+      emit(UserAccountGymError(error: "No internet connection!"));
+      return;
+    }
+
     final response = await userAccountGymService.postUserAccountGym(userBox.get("flexusjwt"), {"gymID": event.gymID});
 
-    if (response.isSuccessful) {
-      emit(UserAccountGymCreated());
-    } else {
+    if (!response.isSuccessful) {
       emit(UserAccountGymError(error: response.error.toString()));
+      return;
     }
+
+    emit(UserAccountGymCreated());
   }
 
   void _onGetUserAccountGym(GetUserAccountGym event, Emitter<UserAccountGymState> emit) async {
     emit(UserAccountGymLoading());
 
+    if (!AppSettings.hasConnection) {
+      emit(UserAccountGymError(error: "No internet connection!"));
+      return;
+    }
+
     final response = await userAccountGymService.getUserAccountGym(userBox.get("flexusjwt"), gymID: event.gymID);
 
-    if (response.isSuccessful) {
-      if (response.body != null) {
-        if (response.body == true) {
-          emit(UserAccountGymLoaded(isExisting: true));
-        } else if (response.body == false) {
-          emit(UserAccountGymLoaded(isExisting: false));
-        } else {
-          emit(UserAccountGymError(error: "User Account Gym Forbidden Value"));
-        }
-      } else {
-        emit(UserAccountGymLoaded(isExisting: false));
-      }
-    } else {
+    if (!response.isSuccessful) {
       emit(UserAccountGymError(error: response.error.toString()));
+      return;
     }
+
+    if (response.body == null) {
+      emit(UserAccountGymLoaded(isExisting: false));
+      return;
+    }
+
+    if (response.body == true) {
+      emit(UserAccountGymLoaded(isExisting: true));
+      return;
+    }
+
+    if (response.body == false) {
+      emit(UserAccountGymLoaded(isExisting: false));
+      return;
+    }
+
+    emit(UserAccountGymError(error: "User Account Gym Forbidden Value"));
   }
 
   void _onDeleteUserAccountGym(DeleteUserAccountGym event, Emitter<UserAccountGymState> emit) async {
+    if (!AppSettings.hasConnection) {
+      emit(UserAccountGymError(error: "No internet connection!"));
+      return;
+    }
+
     final response = await userAccountGymService.deleteUserAccountGym(userBox.get("flexusjwt"), event.gymID);
 
-    if (response.isSuccessful) {
-      emit(UserAccountGymDeleted());
-    } else {
+    if (!response.isSuccessful) {
       emit(UserAccountGymError(error: response.error.toString()));
+      return;
     }
+
+    emit(UserAccountGymDeleted());
   }
 }
